move CountdownTimer to its own component
